requirements.txt
autopep8==1.6.0
pandas==1.3.5
pyyaml==5.4.1
opencv-python==4.5.5.64
tensorboard==2.8.0
scipy==1.7.3
pykitti==0.3.1
parse==1.19.0
iteration-utilities==0.11.0
python-git-info==0.7.1
imageio==2.16.1
pytest==7.1.1
pylint==2.13.0
scikit-image==0.19.2
protobuf==3.20.0
torch_tb_profiler==0.4.0
pytorch-lightning
h5py
hdf5plugin
vtk